# docs/openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "details",
          "request_id"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "BenchRequest": {
        "additionalProperties": false,
        "properties": {
          "allow_physical": {
            "default": true,
            "title": "Allow Physical",
            "type": "boolean"
          },
          "allow_simulated": {
            "default": true,
            "title": "Allow Simulated",
            "type": "boolean"
          },
          "explicit_bench_id": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explicit Bench Id"
          },
          "maximum_wait_seconds": {
            "default": 600,
            "maximum": 86400.0,
            "minimum": 0.0,
            "title": "Maximum Wait Seconds",
            "type": "integer"
          },
          "preferred_labels": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Preferred Labels",
            "type": "object"
          },
          "required_capabilities": {
            "items": {
              "type": "string"
            },
            "title": "Required Capabilities",
            "type": "array",
            "uniqueItems": true
          },
          "required_labels": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Required Labels",
            "type": "object"
          },
          "reservation_duration_seconds": {
            "default": 1800,
            "exclusiveMinimum": 0.0,
            "maximum": 604800.0,
            "title": "Reservation Duration Seconds",
            "type": "integer"
          }
        },
        "title": "BenchRequest",
        "type": "object"
      },
...
      "CiProvider": {
        "enum": [
          "github_actions",
          "gitlab_ci",
          "jenkins",
          "local",
          "unknown"
        ],
        "title": "CiProvider",
        "type": "string"
      },
      "CreateCiSessionRequest": {
        "additionalProperties": false,
        "properties": {
          "actor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor"
          },
          "bench_request": {
            "$ref": "#/components/schemas/BenchRequest"
          },
          "commit_sha": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commit Sha"
          },
          "external_run_id": {
            "maxLength": 500,
            "minLength": 1,
            "title": "External Run Id",
            "type": "string"
          },
          "provider": {
            "$ref": "#/components/schemas/CiProvider",
            "default": "unknown"
          },
          "ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ref"
          },
          "repository": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repository"
          }
        },
        "required": [
          "external_run_id"
        ],
        "title": "CreateCiSessionRequest",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "RunCiWorkflowRequest": {
        "additionalProperties": false,
        "properties": {
          "inputs": {
            "additionalProperties": true,
            "title": "Inputs",
            "type": "object"
          },
          "version": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "workflow_name": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Workflow Name",
            "type": "string"
          }
        },
        "required": [
          "workflow_name"
        ],
        "title": "RunCiWorkflowRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "Phase 4 scoped API token.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Versioned local API for reserving and controlling lab benches.",
    "title": "Lab Platform Agent API",
    "version": "0.5.0-alpha"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/ci/sessions": {
      "post": {
        "operationId": "create_session_api_v1_ci_sessions_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCiSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Session Api V1 Ci Sessions Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create Session",
        "tags": [
          "ci-sessions"
        ]
      }
    },
    "/api/v1/ci/sessions/{session_id}": {
      "get": {
        "operationId": "get_session_api_v1_ci_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Session Api V1 Ci Sessions  Session Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get Session",
        "tags": [
          "ci-sessions"
        ]
      }
    },
...
    "/api/v1/ci/sessions/{session_id}/cancel": {
      "post": {
        "operationId": "cancel_session_api_v1_ci_sessions__session_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Cancel Session Api V1 Ci Sessions  Session Id  Cancel Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Cancel Session",
        "tags": [
          "ci-sessions"
        ]
      }
    },
    "/api/v1/ci/sessions/{session_id}/finalize": {
      "post": {
        "operationId": "finalize_session_api_v1_ci_sessions__session_id__finalize_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Finalize Session Api V1 Ci Sessions  Session Id  Finalize Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Finalize Session",
        "tags": [
          "ci-sessions"
        ]
      }
    },
    "/api/v1/ci/sessions/{session_id}/heartbeat": {
      "post": {
        "operationId": "heartbeat_session_api_v1_ci_sessions__session_id__heartbeat_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Heartbeat Session Api V1 Ci Sessions  Session Id  Heartbeat Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Heartbeat Session",
        "tags": [
          "ci-sessions"
        ]
      }
    },
    "/api/v1/ci/sessions/{session_id}/run": {
      "post": {
        "operationId": "run_session_workflow_api_v1_ci_sessions__session_id__run_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCiWorkflowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Run Session Workflow Api V1 Ci Sessions  Session Id  Run Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Run Session Workflow",
        "tags": [
          "ci-sessions"
        ]
      }
    },
...
  }
}